requirements.txt
fastapi
uvicorn[standard]
google-cloud-firestore
google-cloud-aiplatform
feedparser
sendgrid
requests
